session: Implement logic to load a rough desktop session
As we dont yet currently provide a panel or window manager of our own,
we launch mutter and tint2. The session program does not exit until the
window manager has quit, ensuring there is no premature abortion of the
desktop session.

Also note, as we have no application launchers yet (due to no panel of
our own) we also have to launch a DESKTOP_EXTRA component, also known
in some circles as 'gnome-terminal'.