panel: Use X11 struts to ensure panel size is counted by WM
This is slightly bad, but presently the best tactic to use. Given that
libwnck is X11 dependent, it is ok for us to use another piece of
X11 dependent code, until we replace the tasklist with something
more generic.

This means we need to support plugins via libwnck, and runtime checks
to determine if we're using Wayland or X11, and alter our behaviour
accordingly. Eventual aim is to drop hard-coded requirement on X.