menu-window: Allow searching by description
Sometimes a user may have a rough idea what to search for, but not
actually know the application name. For this we can simply search
the description of installed applications as well as searching
by application name.

Also fixed a potential issue where a description is NULL, which would
have resulted in a segfault.