wm/plugin: Zoom in the desktop from center
Now that we update our wallpaper from the settings, it is now possible
to use an effect when changing it.
Note that this only applies a zoom to the *new* wallpaper, it merely
removes the old one, leaving what is probably right now quite an
annoying white screen before it does so.
This will be rectified in the future when we switch to MetaBackground
actors, using load_async, etc.