wm/plugin: Workaround gnome-control-center hackery
When gnome-control-center sets a colour for a background, it puts the
setting into wallpaper mode, using a noise pixmap. The desired behaviour
is obviously to show a colour, but this is GNOME specific behaviour.

When we detect this behaviour, we simply workaround it, display the colour,
and do not display the noise pixmap. We *may* support the noise pixmap
and colour combination in the future.