applets/menu-window: Visual fixups
We now use the symbolic shutdown icon, and a search entry in the
top of the menu window, similar to Chrome OS's menu window.
Icons have been enlarged, now that our window does not take up as
much space. The default style provided by budgie-panel's css
forces the GtkSearchEntry style to blend in with the popover.

As and when compatible themes appear, this can be changed.