Curly braces everywhere.
Single line if-statements without curly braces are generally frowned
upon these days, as they have the potential to introduce fairly
disastrous consequences later on in the life of budgie-desktop..

Coding style for budgie-desktop now *requires* every if statement
to be wrapped with curly braces. The last thing we want is some
piece of code executing when its not meant to.. or making assumptions.