Force some styling decisions - drop CSD hack usage
We now explicitly draw ourselves and have a fixed border width of 11
to draw within, to counter the many broken themes out in the world.
We can no longer use CSDs to draw ourselves as they're unrealiable
and broke/changed between 3.10 and 3.13 too many times.
In future we'll do window shaping and avoid most of the GTK weirdness
altogether - and composite within budgie-wm.
We also now ensure we have a border-radius of 6px, basically because
no themes support Budgie and it looks crap everywhere.