Strip back license headers to sensible sizes
We ship the entirety of budgie-desktop under the GPLv2 license - however
we don't need to include that much text when we also ship the license file
itself. This commit serves as record as such, and you should indeed consult
said LICENSE file and know that GPLv2 applies to budgie-desktop as a whole,
including build files and extranous data files such as CSS supplements.