budgie-popover: Do not unmap/unrealize the widget
In the past we unrealize()d the widget, which has the effect of freeing
all related Gdk resources. This in turn makes it much slower for the
popover to appear in the first place, as the resources must be regained.
We now ensure to track our own known dimensions and use those so that
we have no need to rebuild ourselves.