panel: Rewrite BudgiePopover in Vala
This is a complete rewrite of the BudgiePopover widget in Vala, and is far
superior to the previous C iteration. This particular implementation correctly
handles the various cases for grabs, actively steals the grab back when
loosing it to a popup menu (entries), is quite a lot cleaner and can handle
multiple positions.

Note this version has some special workarounds in place to enable integration
with the MenuWindow class, as inheriting from Vala classes is somewhat
difficult to achieve (superconstructor doesn't work correctly) - that said
in the future the entirety of Budgie will be rewritten, so it is no big
issue.