panel: Replace old panel with new Vala panel
This is only partly implemented right now, we have a base panel that
serves much of the same functionality as the previous one, but currently
has no applets (TODO)

This panel is much better than the previous iteration, and far more
maintainable too. (Win) This panel handles vertical sizing correctly
and will ensure content is restricted within the bounds of the
panels width. Positioning is correctly implemented, though specialised
applet support will be required for a vertical orientation, as
"shit will break (TM)"

Note to readers: This does not yet incorporate GSettings, nor is it
usable yet by a single soul (empty panel) and does not yet include
the GNOME Panel theme integration, due to aforementioned lack of
applets.