panel: Re-introduce auto-darkening of panel
The panel will once again darken itself (the usual style) when a window
becomes maximised. Now we're using Vala it makes sense to actually iterate
the list of windows and just go ahead and check them all each time something
changes to ensure we're consistent.

NOTE: This doesn't yet support changing of workspaces and that's likely
where it will break, but so does Budgie in general.