applets/ClockApplet: Use low priority callback for clock update
This no doubt will come at a slight cost of accuracy, however there is
really no need right now to display at precisely 1s an updated clock,
especially as it doesn't even show seconds.
Its a complete waste of power, and for this reason we use a low priority
callback grouped with others to run in blocks, ensuring we're not wasting
precious cpu time.