session: Ensure CSD hack header is invisible
To ensure the BudgieSessionDialog is both rounded and has a shadow, we take
advantage of CSD features within Gtk. However with some Gtk changes, and even
some Adwaita changes, the empty box is now visible on Gtk 3.12. This commit
simply ensures that the header widget is transparent and does not show
as a white line.