Inroduce Panel Editor
Introduces a somewhat unoptimized way of editing the panel layout. Note there
is currently no shortcut for accessing this editor, so it is still in effect
hidden from the user.

The Panel Editor is simply the beginnings of a new management UI to allow
dynamic positioning, adding and removing of applets. It has a lot of work
to go through before it will be considered stable.