panel: Add initial GNOME Panel theme support
This works on an extremely simply premise which thus far seems to work quite
well. The vast majority of themes that tailor to GNOME Panel explicitly use
"PanelToplevel" and "PanelApplet" types to do so. We simply take advantage
of this by using a PanelToplevel (inherits from BudgiePanel) and wrap all
applets in a PanelApplet class (an empty Gtk.Bin)

This setting is only read once, at panel start time. You still need to restart
the panel for the setting to take effect.