applets/budgie-menu: Reload menu tree upon changes
This ensures we keep our tree data up to date with what is actually present
in the gnome-applications.menu directory whenever it is changed, i.e. when
a package is removed/updated/reinstalled, etc. This also stops the panel from
crashing any time .desktop files are modified, as well as the nasty (null)
labels previously seen.

(Or, this should've been re-implemented back at the vala rewrite..)