Introduce legacy X11 tray.
This is reliant on NaTray, nabbed from GNOME Panel.I have no intention in
maintaining this longer than I have to, and firmly believe xembed should
be taken outside, tied to a tree and have both legs shot with nail guns
while slowly being tortured to the sound of failed Disney artists.

The main rationale in including this right now is the lack of agreed
"indicator" spec between major desktops, meaning we will support our
own applets primarily, xembed cruft as a last option and in the future,
appindicator.

People need to be able to access programs that still use the tray,
such as "nm-applet", xchat, etc. Now I've explained why this is here,
please, enjoy the bugs.