applets/status: Introduce volume control and unified popover
The unified popover (or "Aggregate menu") enables the quick changing of, and
access to, common desktop tasks. Right now this involves a link to the
settings, interactive control of volume, as well as an option to end the
session. This will be expanded in the future to support networking and
devices.