Cosmetics: Initial shadow support, and native popover integration
With these changes we currently support a panel shadow (via a gsettings key)
only at the *bottom* of the screen, this will be revised in future. Also with
these changes comes "native" popover integration, which essentially means
we do some kung-fu to take advantage of the Gtk theme for GtkPopover, meaning
we don't need tons of BudgiePopover styling.

Kudos to the Wingpanel developers for the panel shadow concept.

Also note this commit currently breaks GTK3.10 builds due to the usage of
GtkPopover, which will be resolved in a follow up commit.