applets/icontasklist: Implement application pinning
It is now possible via GSettings (pinned-launchers) to permanently place items
on the panel at the start of the IconTasklist applet. Once activated they will
launch the application, and as and when the application has been launched,
they will behave identically to the other launcher items, with the actions
being the same, i.e minimize, restore, right click exposes window menu.

Note there is currently no UI control for adding these launchers yet and it
is still largely an experimental feature.