panel: Only resize when we're *absolutely* required to
Prior to this commit the size_allocate routine saw us using anywhere
up to 50% of the CPU in some cases, which is absolutely ludicrous. The
desktop as a whole will now be far more responsive as the panel won't
be crunching the CPU and attempting to resize/loop/reload images
constantly.
I personally apologise for my failure to address this sooner.