panel: Ensure context menu works always on Ubuntu
This *appears* to be a problem existing in GTK3.10, however it never actually
manifested on Evolve OS using GTK 3.10. The assumption would be there is an
issue with one of the Ubuntu GTK patches.

This ensures all events are available to the toplevel GtkWindow so that the
right click (context) menu will work regardless of where the user right click,
which no doubt has so far been much to the annoyance of our Ubuntu audience.