panel: Overhaul the existing editor UI
Long story short its now a lot more sexy and usable. It's evolving (punny)
into a new Settings system, which is designed to remove the need on external
tools within Budgie on things like gnome-tweak-tool for simple operations
such as changing the desktop theme.

Note that less stack-abuse is present now, and we rely on a dialog to show
the "add applet" view.