wm: Introduce the new WM!!!
So yeah, this needed to be done for a really really feckin long time.
Most of it is done, opening the door up to much improvement now in the
future. Some bits of it are lacking, like the alt+tab, but we can fix
it, we have the technology.

Remnants of the old plugin are currently utilised via legacy.c, and
will be purged in the near future to ensure a clean codebase. And no
damn 2-space indentation! But yeah, the original was a modification of
a demonstration plugin. That's what you get for being lazy.

Note: We've pushed the minimum Mutter API to 3.14 for the moment, until
some 3.12 testing can happen, and we'll see how realistic it is to support
both platforms. For the minute, its 3.14 because Evolve OS is on 3.14,
and that's priority. Help gladly accepted in checking 3.12 out..