widgets: Introduce "Budgie.Animation"
In short, this is a Clutter inspired implicit(ish) animation which works
by modifying GObject properties. As we're a Gtk desktop, it's kinda hard
to actually integrate Clutter and stay in line with our own goals. As such
we'll use our own tiny animation helper until such time as Gsk is ready
for "prime time"