widgets: Introduce "PaperButton"
Ok, so its not paper, and its not quite material design. But it has a ripple
so whatever. Looks kinda shmexy. In future we'll need to *not* use a subclass
of GtkToggleButton as that is pretty much cheating, rather we need to add
specific interfaces and an InkManager (Ugh, Java much?) - to quickly plug
in the rendering behaviour into "normal" widgets. Instead of rewriting them.