wm: Trial inclusion of force unredirect configuration
On certain devices, disabling unredirect of windows can result in performance
improvements in full screen mode, i.e. when gaming. However, in certain
configurations this can actually result in a negative user experience, i.e.
through excessive tearing. As such, we expose this as an option, and not a
default, for those looking for the best performance whilst gaming.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>