rundialog: Quit after losing focus
The rundialog isn't a real "app" as such, it is simply a callable utility
that when interacted with, launches another application. As it needs no
persistent or reserved screen estate, we should simply close it when the
window loses focus, as it is very cheap to just invoke it again with ALT+F2.

This addresses issue #494.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>