applets/status: Remove the long despised power-off symbol
The existing icon was a source of confusion as it only ever opened Raven,
which in turn had a power icon, which then invokes a dialog.. As of now,
we also have a User Menu applet, which provides the appropriate controls
for the session, in a much more intuitive fashion.

This removal is appropriate due to the now redundant fashion of the icon.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>