applets/status: Ensure power menu doesn't show when batteries are hidden
Previously we accidently called a show_all, which meant the power applet
was inadvertently visible even when it explicitly removed all it's children,
i.e. sync invoked before visible and hide() invoked.

Now we ensure that we're no longer in the popover menu chain by forcibly
making sure the eventbox is hidden when it is supposed to be.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>