panel: Fix off-by-one that randomized placement of panel applets
We incorrectly set the tail position for newly added applets to the size of
the list, which would mean the position would be larger than the 0-indexed
position child property within the box layout.

Given the invalid nature, this would have the equivalent of adding to the
parent GtkBox area with -1, i.e. next free slot. Subsequent load additions
of applets might then result in position conflicts, which would cause even
further randomization and applets swapping, notably the last 2 applets of
the panel, as recently seen with the introduction of the Raven Trigger.

This is a ridiculous derp by me and I'll need to figure out a migration system
to help undo this, as conflicting position keys may exist in their GSettings
configuration for the panel.

To further mitigate future issues, we now ensure in our initially loaded
loop check that first *all* applets are correctly parented in their region
group, and *then* apply repositioning, to avoid race conditions and ensure
that all children are loaded before organising them within the panel.

This resolves issue #503, and undoubtedly many more.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>