panel: Don't lie about the visible panel size
We always, in every case, reserve 5px for the shadow region, regardless of
whether it's visible or not. So if the intended_size is 39px, 5px is allocated
for the shadow, leaving 34px for the visible region (the main panel layout).
As such, this is the size we should inform the children of, which also fixes
issues with the aspect ratio in the Icon Tasklist Applet.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>