applets/icon-tasklist: Remove now-unusable "large-icons" setting
In the past this option was required for toggling between normal and large
icon sizes. However, as Budgie has advanced we have added automatic icon
size scaling, as well as enforcing an aspect ratio into this applet.

As such, the large-icons setting looks entirely derpy and breaks the visual
consistency/usability of the panel.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>