applets/budgie-menu: Enable rolling over categories to open them
This enables a simple rollover ability with the mouse to switch between
the various categories, as opposed to manually clicking on them. Note this
isn't the default, because some people will find this problematic, depending
on their mouse (especially touchpads). However the option is now there for
those who do need it in their workflow.

This resolves issue #426.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>