panel: Fix the locally supported user extension path
Per issue #385, the current user path is invalid, pointing to "modules" and
not "plugins" (harkening back to older Budgie). We now use the "plugins"
local directory by default, and only add the "modules" directory if it exists
at startup, emitting a warning that the path will be *removed* in the next
major version of Budgie.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>