Add new schema for the migration-level
As part of issue #555, we need to unblock the migration path in Budgie to
account for the changes in the applet configurations. We will begin by
tracking the migration-level of the installation, currently set to 0, and
if the migration-level is less than the current minimum required migration
level (1), we'll perform migratory actions on *existing* panels.

In future this will allow for automatic deprecation/replacements of applets
without breaking the upgrade experience for users.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>