panel: Finalize the Version 1 Migration Path
This now resolves issue #555, which in turn unblocks the Budgie 10.2.7
release. In the instance that migratory requirements are not met, we'll
add the 2 new applets to compensate. This is basically everyone currently
running the 10.2.6 release. We put them at the end of the panel for existing
configurations, in order to retain simplicity.

For new configurations, i.e. fresh installs, no migratory action is taken
and we use the defaults. In all instances we immediately update the stored
migratory level on initialisation to ensure there are no erronous subsequent
attempts to perform costly migration.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>