wm/shim: Add proxying for the org.gnome.Shell ShowOSD function
This provides the appropriate proxying functionality to enable Budgie to
support the GNOME Shell OSD API for GNOME Settings Daemon. As we cannot
implement GTK+ functionality in Budgie WM itself, we'll forward this to
an upcoming GTK+ component to manage OSD handling outside the WM in
a nice controlled fashion.

This is part of issue #261

Signed-off-by: Ikey Doherty <ikey@solus-project.com>