daemon: Add main runtime manager
The main manager handles the main lifecycle of budgie-daemon, providing
global state updates for the theming + GTK CssProvider according to the
estabilished Budgie settings and HighConstrast config.

It also provides the initial GNOME Session integration to ensure this can
become an early-init component of Budgie prior to the load of the panel,
and appropriately be tracked in the session.

This is ongoing work as part of issue #261

Signed-off-by: Ikey Doherty <ikey@solus-project.com>