Move the EndSessionDialog inside budgie-daemon
Budgie Panel was never a good location to have the EndSessionDialog, however
in the past we never could justify a separate daemon. However, with the
introduction of the OSD functionality in budgie-daemon, along with upcoming
labeller work, and a future replacement for GNOME Settings Daemon, this is
the ideal location to have this dialog.

A key benefit being that if the panel is crashed for whatever reason, it
can continue to live on inside the budgie-daemon at no risk.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>