applets/icon-tasklist: Use case insensitive query always
A recent change in Google Chrome apps, using different profile fields such
as "-Default.desktop", exposed a bug in the icon tasklist in that it tries
to do a case sensitive lookup for the .desktop file.

Unfortunately we'd already stored them with lower-case to reduce the risk
of missing lookups. You live, you learn. We now *only* do case-insensitive
matching in icon tasklist which resolves issue #566 and allows proper
Chrome applications to pin once again.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>