applets/status: Implement a volume popover on the volume applet
While this may at first appear redundant, it has to be remembered that
Budgie is interacted with in a few different ways. Whilst mouse users can
scroll on the applet to adjust the volume, this isn't always easy, nor is
it always intuitive/discoverable, when using a touchpad on a laptop.

Another issue is that with recent changes to the "indicator-like" applets,
the volume icon sticks out like a sore thumb as it is the only one to do
absolutely nothing when clicked.

Therefore to increase consistency and usability, we provide a popover widget
when left clicking the volume icon, with an adjustment scale and buttons similar
to those found in media players, to make it very simple to control the system
volume on the current stream.

As part of this change, the duplicity in StatusApplet.vala was partially fixed
by using factory methods for dealing with the various child popovers.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>