daemon/endsession: Correctly handle changes to compositing state
Per issue #581, the End Session Dialog no longer has shadows, or indeed,
transparency, due to setting a null RGBA visual during start up. This is
because we're starting up before the WM is fully initialised and we don't
have an RGBA visual, due to a lack of composite extension.

We now correctly track the changes and install the RGBA visual the first
chance we get.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>