applets/status: Correctly port to the upower 0.99 API
This is a serious oversight which means the remove signal would always result
in a segfault, as we're trying to treat a string as a gobject. Sorta figures
it would go bang.

This reenforces the case for using C even further by not masking issues under
a vapi file. This resolves issue #594.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>