fix: HeaderWidget wasn't notifying HeaderExpander of changes to 'expanded' prop.
RavenExpander/HeaderWidget was not passing to the HeaderExpander button changes to
the expanded property. When adding a widget that starts in a collapsed
state, the HeaderExpander would not be updated accordingly and would
remain in its default state, expanded = true, with the downwards arrow
icon. This bug occurrs, for instance, with Spotify.