panel: Chain methods to fix GtkBox allocation spam every half second
In newer versions of GTK (Notably, 3.20 onwards) we're spammed every second
or so with debug messages regarding the size allocation. In reality, this
is ticked by the clock update and any other applets requiring refresh.

As such we now chain the method to the base methods and ensure that the
internal state machine is updated accordingly, saving us from ridiculous
levels of spam in the journal and a complete inability to debug the panel.

This resolves issue #532 and it's duplicates.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>