plugin: All applets now extend GtkEventBox and not GtkBin
As seen in issue #574, we have various problems with certain applet types
due to extending from the primitive GtkBin. In this particular instance we
have no child GdkWindow, which is why the drawing system becomes bizarely
bugged and results in rendering in random locations.

All applets should have at least one top level GdkWindow that is not part
of the GtkWindow's GdkWindow, to ensure separate context. As a part of this
change, all third-party applets should be rebuilt against the new Budgie
due to the ABI break introduced here.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>