ibus: Ensure we correctly init only after we know we're connected
This ensures that ibus types are actually registered so that serialization
works. If we attempt to ibus_init *before* this, newer versions of ibus
will be d-bus activated, and spoil exactly how we control the workflow.

TL;DR We need to work with old and new alike, and not have ibus randomly
spawn before we tell it to with a tray icon.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>