applets/icon-tasklist: Ensure we really reload appsystem
Previous our GAppInfoMonitor was unreffed immediately, meaning we never
actually got the changed signal. So any time a new application was installed,
it wouldn't actually show up in the menu. Audacious was the most recent
example of this.

We now keep a life time reference to the GAppInfoMonitor, and implement a
basic lazy cache system. Essentially we mark the cache as invalid, and the
next time a desktop file is due to be checked, i.e. to query the desktop file
for a newly launched window, we rebuild the cache and mark it valid again.

The net benefit is that you can install an app, and immediately have access
to the pinning functionality.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>