applets/icon-tasklist: Implement Exec based mapping as the ultimate fallback
If every other form of lookup has now failed, we try to match on the TryExec
and Exec lines, sanitized, cleaned and basenamed. This won't always work,
but for the rare few cases it may just be enough to be complete.

This specifically fixes #627 whereby GIMP isn't pinnable as the window class
is "gimp-2.8" and unaccounted for anywhere (gimp.desktop)

Also, I'm definitely going to hell for this.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>