raven: Ensure all cursor themes are loaded
Previously we made a ridiculous assumption that icons and cursors were
somehow mutually exclusive. The idea of the index.theme parsing was
simply to test that it was a valid *icon theme*.

However, we then only tested for the cursors directory if it definitely
wasn't an icon theme. That meant that themes shipping both cursors and
icons in one directory would never show the cursor themes.

This fixes #623.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>