Always restore focus to windows after closing popovers
Drawing some inspiration from the downstream work in Wingpanel, which uses
the Budgie popover concept, we store and restore the focused window in the
window manager around expansion changes in the panel.

Essentially, before first expanding (pre-popovers) we store the currently
focused window, and when we shrink back to normal (no more popovers) we ask
the window manager to restore focus to the window that last had it before
we went and got all funky on their display.

This fixes #530

Signed-off-by: Ikey Doherty <ikey@solus-project.com>