Make all Raven Mpris clients, e.g. Spotify, start expanded
Raven was deciding whether to start Mpris clients expanded based on the property
'CanQuit' of the MediaPlayer2 dbus interface. But according to the interface
spec, this property should be used to know if we can close the application
through its Mpris client, e.g. through a close button in the client's UI.
This commit simply makes all Mpris clients start expanded.