raven: Ask the user for confirmation before removing a panel
Currently it's very easy to remove a panel, and for the uninitiated the
first thing they might do is accidentally delete the top panel thinking
they're about to remove the bottom panel.

We now show a confirmation popover when clicking the removal (-) button
to ensure accidents *don't* happen.

This is another part of issue #289.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>