Dynamically update geometry in response to HiDPI scaling changes
The user, or the system, may decide to enforce a new scaling factor on the
UI at any given time. This may be in response to switching to a 4K output
or similar.

As such, when we detect a hidpi change we should re-assert our struts and
position. This change has been backported from the rewrite to enable it
to go into the minor release, and fixes #679.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>