theme: Backport BudgieThemeManager from budgie-next
The BudgieThemeManager handles the monotonous task of ensuring all Budgie
processes follow the correct theming. This includes switching between the
normal and high-contrast variants, and monitoring the changes in the gtk
theme to dynamically apply the correct theme.

Care is taken to also select the correct theme version from the built in
resources.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>