theme: Build resources as static archive and then link into the .so
The theme resources compilation emits many spammy warnings that are just
totally unavoidable, so we've always built that portion without any extra
CFLAGS. This portion does nothing more than register the resources, and
the only optimisation that would help would be LTO.

However, our shared library itself contains real code now, so should then
follow the global CFLAGS, especially if we want to ensure the code is safe
and clean. We now build the resources as a static archive, and then link it
into the final libbudgie-theme library.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>