fix: MprisWidget was not notifying its clients of size changes
Before Raven is opened a first time, any new ClientWidget created will be
assigned a default width of 250. But, once Raven was opened and its size was
allocated, the new width would not be passed to the ClientWidgets, causing their
cover art to have an incorrect size. With this commit, MprisWidget will notify
its clients whenever its size allocation changes.