applets/keyboard-layout: Fallback to locale based language detection
In the display string, for the zh_CN locale, we did not manage to retrieve
a language name from libgnome-desktop. This resulted in rendering an odd
label for libpinyin ibus engine, i.e. "(null) (libpinyin)".

When we fail to get a language based on the language code, we should fall
back to locale detection, which should render in the appropriate locale
too.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>