applets/icon-tasklist: Fix the Spinny Wheel Of Doucheness
This change ensures we don't accidently create a new startup ID, which
was only discovered after digging into the GTK source code to figure
out that a "get" function is in fact a "set, send, return", which in
turn is then reused inside th GDesktopAppInfo implementation...

So now we remove the hacks, use a per-button GdkAppLaunchContext, and
correctly "turn off" the notify via Gdk to stop the spinning wheel
when launching applications.

In short, startup notify support is properly implemented in the
panel now, thanks to some investigative work done for the Brisk Menu.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>