rundialog: Only exit once the application has launched
We now monitor dbus names to see when the application actually comes
up, as opposed to niavely immediately quiting. This does mean some
quite nasty workarounds as we now have to monitor names on the bus
compared with the .desktop files.

For "normal" applications we obey the startup notification spec and
wait for them to fail or succeed, and quit then. It's also quite likely
that we'll have to add functionality to expire the rundialog after a
certain amount of time regardless of the call, in case dbus activation
failed (cuz it succeeds regardless) - and to ensure we show again when
we are app-activated.

This resolves issue #658.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>