wm: Add basic animation state tracking for MetaWindowActor
Instead of relying on complex lists to manage the various animation types,
we store a global hashtable which maps a window to a bitmask. When all
masks have been removed from a window, we can then remove that mapping
from our hashmap.

This allows us to gracefully terminate any animation state for a window
as well as support multiple concurrent animation states. Once we add
the actual tweens themselves we can take care of removing any lasting
animation effects (i.e. restoring actor state) and removal of the
clutter effects.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>