Revert "wm: Restore C version to continue hacking on this" (issue #800)
This reverts commit 5096925b5b92dbb93cff755073a67912243938a2.

C WM is no longer a target. I'm not going to waste my time doing a complete
rewrite, only to do another rewrite for Budgie 11. This is just a poor time
investment. Instead we'll try to plug something in and clean it all up to
make it last for a while.