daemon: Add a new --replace mechanism
This makes replacements of the daemon deliberate, not accidental.
Additionally we make it very clear to the user that we've done a replace.
Most importantly, we don't end up with "spare" daemon processes lying around
that have a main gtk loop but no d-bus, thus we quit the processes there
and then.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>