tabswitcher: Use a horizontal icon list
This is a far more streamlined view as it makes use of the horizontal
space, instead of masses of vertical space. This new method makes use
of a horizontal GtkFlowBox, so if necessary it can reflow onto a new
row.

As a further space saver, we add a single text label at the bottom of
the switcher as opposed to each item having its own.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>