Add network manager to the spamlist for notifications
These notifications appear on screen, as well as having a permanent place
in the panel by way of the nm-applet status icon. It has long been seen
that upon start up we end up with an undesirable spam notification clone
into Raven to tell us that we connected to the network.

On a side note, this fixes the "red icon on login" issue..

Signed-off-by: Ikey Doherty <ikey@solus-project.com>