panel: Fix troublesome popovers once and for all.
This commit fixes two issues: GTK 3.22 animations (doubling) and, finally,
the glitchy menu on the bottom panel. Granted, the code is especially evil,
but it actually is enough to trick GTK+ past it's own trip-ups and place
the popover in the right position every time.

This fixes 292 and it fixes 809 properly too.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>