applets/budgie-menu: Remove the antique "app-scores" concept
This one is another theft/adaptation from Brisk Menu, in which we dynamically
compute scores for a given entry in relation to a search term. Assuming it's
relevant, we'll display in the given order.

This allows terms to match and higher up in the list depending on how close
to the term it is, the contents, how much of a string is matched, etc. This
now removes the previous behaviour that some dubbed as "learning", by instead
doing the Right Thing™.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>