applets/budgie-menu: Behave better when using compact modes
When headers aren't visible, we should sort only based on the application
name in the "All" view, as the headers are supposed to separate the blocks.
This feels more consistent on both methods.

When in "super compact" mode, i.e. no headers, and sidebar categores are
hidden, we prevent any duplicate entries from ever appearing, as there is
no method to actually change category anyway.

This fixes #512.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>