plugin: Introduce the new BudgiePopover type
The BudgiePopover is intended to be a complete replacement for the
GtkPopover use within Budgie Panel applets. A GtkPopover is not actually
a top-level window, and must be contained within another toplevel window.

Consequently, Budgie has a large series of compounded hacks to dynamically
resize the main panel window to the full size of the display, and include
the popover within that rendering region.

Unfortunately, this is incredibly buggy (and inefficient), and is virtually
impossible to use with the bottom panel configuration, as well as violating
the expectations of what a dock window *can* and *should* do, i.e. our panel
window is set to "always-on-top", which is plain and simple: wrong.

The new BudgiePopover is a native window that appears to the window manager
as a simple popup menu. Internally it has much in the way of sizing and
positioning logic, as well as rendering code, to emulate the appearance
of a GtkPopover. However it doesn't use any GtkPopover code and is the
spiritual successor of the original BudgiePopover, before I temporarily
lost my mind and used Vala and tried to make Budgie conform.

The new BudgiePopover will be used in conjunction with a newly implemented
BudgiePopoverManager, to provide the same menubar-like roll-over activation
functionality that existed before, but with the benefits of using a proper
X11 grab, and being native top level windows.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>